backend/requirements.txt
Flask>=2.3,<3
flask-cors>=3.1.0
python-dotenv>=1.0.0
